+++ src/vscode/uriHandler/FabricUriHandler.ts
@@ -1,4 +1,7 @@
+import { Helper } from '@utils/Helper';
 import * as vscode from 'vscode';
+import { FABRIC_SCHEME } from '../filesystemProvider/FabricFileSystemProvider';
+import { ThisExtension } from '../../ThisExtension';
 
 export class FabricUriHandler implements vscode.UriHandler {
 
@@ -9,9 +12,29 @@ export class FabricUriHandler implements vscode.UriHandler {
 	}
 
 	public async handleUri(uri: vscode.Uri) {
-		const params = new URLSearchParams(uri.fragment);
-		const workspace = params.get("workspace");
+		//https://vscode.dev/+GerhardBrueckl.fabricstudio/open?workspaceId=MyWorkspace
+		ThisExtension.Logger.logInfo(`Uri-Handler invoked: ${uri.toString()}`);
+		
+		switch (uri.path) {
+			case "/open":
+				this.openFromUri(uri);
+				break;
+			default:
+				vscode.window.showErrorMessage(`Unknown command: ${uri.path}`);
+				break;
+		}
+	}
 
-		vscode.window.showInformationMessage("Workspace: " + workspace);
+	public openFromUri(uri: vscode.Uri): void {
+		const params = new URLSearchParams(uri.query);
+		const workspaceId = params.get("workspaceId");
+
+		if (workspaceId) {
+			Helper.addToWorkspace(vscode.Uri.parse(`${FABRIC_SCHEME}:/workspaces/${workspaceId}`), `Fabric Workspace`, true, true);
+		}
+		else {
+			vscode.window.showErrorMessage("No workspaceId provided!");
+			return;
+		}
 	}
 }